+++ client/src/api/user.ts
@@ -43,3 +43,14 @@ export const PostHelpPw = async (email: string) => {
 
   return response;
 };
+
+export const PatchProfile = async (userId: number, formData: FormData, token: string) => {
+  const response = await axios.patch(`${URL}/user/${userId}`, formData, {
+    headers: {
+      "Content-Type": "multipart/form-data",
+      Authorization: `${token}`,
+    },
+  });
+
+  return response;
+};
